Trading/GenX_FX_MASTER/A6_9V_Organization/requirements.txt
# Core ML and Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Technical Analysis and Financial Data
ta-lib>=0.4.26
yfinance>=0.2.18
mplfinance>=0.12.9b7

# Machine Learning Extensions
joblib>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing
aiohttp>=3.8.0
aiofiles>=23.0.0
asyncio-mqtt>=0.16.0

# Excel and CSV Processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Configuration and Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# Async and Concurrency
asyncio>=3.4.3
uvloop>=0.17.0

# Logging and Monitoring
loguru>=0.7.0
psutil>=5.9.0

# API and WebSocket Support
websockets>=11.0.0
requests>=2.31.0

# Math and Statistics
statsmodels>=0.14.0
pykalman>=0.9.5

# Optional: FXCM Trading (if using real broker)
# fxcmpy>=1.3.0

# Development and Testing
black>=23.0.0
flake8>=6.0.0

# Time Series Analysis
arch>=6.2.0
pmdarima>=2.0.0

# Risk Management
# pyfolio>=0.9.2  # Not compatible with Python 3.13
# empyrical>=0.5.5  # Not compatible with Python 3.13

# Notification Support (Optional)
# python-telegram-bot>=20.0
# slack-sdk>=3.21.0
fastapi==0.103.2
pybit==5.11.0
uvicorn==0.22.0
python-jose==3.4.0
pydantic-settings==2.0.3
